Fix the issue of shipping address when the session expires
Fix the issue of shipping address when the customer has a new account created from checkout page
