+++ checkout-multishipping/view/frontend/templates/checkout/link.phtml
@@ -8,7 +8,7 @@
 <a class="action multicheckout"
     href="<?= $block->escapeUrl($block->getCheckoutUrl()) ?>"
     data-mage-init='{
-        "RefactoredGroup_AutoFflCheckoutMultiShipping/js/checkout/handle-checkout-multiple-address-button": {}
+        "RefactoredGroup_AutoFflCheckoutMultiShipping/js/checkout/proceed-to-checkout-multiple-addresses": {}
     }'
 >
     <span><?= $block->escapeHtml(__('Check Out with Multiple Addresses')) ?></span>
