+++ checkout-multishipping/view/frontend/web/js/checkout/proceed-to-checkout-multiple-addresses.js
@@ -1,0 +1,22 @@
+/**
+ * Copyright © Refactored Group (https://www.refactored.group)
+ * @copyright Copyright © 2022. All rights reserved.
+ */
+
+define([
+    'jquery',
+    'Magento_Checkout/js/checkout-data'
+], function ($, checkoutData) {
+    'use strict';
+
+    return function (config, element) {
+        /**
+         * If the "Check Out with Multiple Addresses" link is clicked,
+         * set the proceedToCheckoutWithMultipleAddresses value to true.
+         */
+        $(element).on('click', function (event) {
+            checkoutData.setProceedToCheckoutWithMultipleAddresses(true);
+        });
+
+    };
+});
