+++ checkout-multishipping/view/frontend/web/js/checkout/style-shipping-address-table.js
@@ -1,0 +1,24 @@
+/**
+ * Copyright © Refactored Group (https://www.refactored.group)
+ * @copyright Copyright © 2022. All rights reserved.
+ */
+define([
+    'jquery',
+    'Magento_Checkout/js/checkout-data'
+], function ($, checkoutData) {
+    'use strict';
+
+    return function (config, element) {
+        /**
+         * If "Check Out with Multiple Addresses" is clicked, ignore
+         */
+        if (checkoutData.isProceedToCheckoutWithMultipleAddresses()) {
+            return;
+        }
+
+        /**
+         * Add new class name to enable custom styling
+         */
+        $(element).addClass('ffl-shipping-address-table');
+    };
+});
