+++ core/view/frontend/web/css/source/_module.less
@@ -300,7 +300,7 @@ select.dealers-modal-button {
 }
 
 /**
- * Style the .ffl-grid table
+ * Style the .ffl-shipping-address-table table
  * 
  * This initially sets the border of td element to none
  * and the div containing the address input field display to none.
@@ -309,7 +309,7 @@ select.dealers-modal-button {
  * and resets the above styles.
  */
 @media all and (min-width: @screen__s) {
-    table.ffl-grid {
+    table.ffl-shipping-address-table {
         tbody {
             tr {
                 td {
