+++ core/view/frontend/web/js/cart/dealers-popup.js
@@ -107,11 +107,15 @@ define([
                     /**
                      * Assign the address to the elements on the Cart page.
                      * 
-                     * First, it checks if the "Proceed to Checkout" button is clicked.
+                     * First, it checks if the "Check Out with Multiple Addresses" is clicked.
                      */
-                    if (checkoutData.isFromCheckoutPage() &&
-                        checkoutData.isFflProceedToCheckoutButtonPressed()
-                    ) {
+                    if (checkoutData.isProceedToCheckoutWithMultipleAddresses()) {
+                        /**
+                         * Default to setting the dealer address to individual text input.
+                         */
+                        dealerButton().dealerAddress[self.currentFflItemId()](parsedResult.name);
+                        dealerButton().dealerAddressId[self.currentFflItemId()](parsedResult.id);
+                    } else {
                         /**
                          * If true, fetch the row index of FFL items from localStorage.
                          * Then iterate through these items and assign the value of the ID
@@ -123,12 +127,6 @@ define([
                                 dealerButton().dealerAddressId[element](parsedResult.id);
                             });
                         }
-                    } else {
-                        /**
-                         * Otherwise, default to setting the dealer address to individual text input.
-                         */
-                        dealerButton().dealerAddress[self.currentFflItemId()](parsedResult.name);
-                        dealerButton().dealerAddressId[self.currentFflItemId()](parsedResult.id);
                     }
 
                     // If we are on the multi-shipping checkout shipping page, reload
